requirements.txt
fastapi
uvicorn
pydantic
python-multipart
pyjwt[crypto]
pytest